Initial Commit
Has all previous iterations and an example hex reader example from someone else..

Currently can choose either raw mode (character by character is sent to the serial) or line mode.  It pre-populates the COM ports and checks every second for updates if not connected to one., but I think it only works on windows at the moment (not hard to fix this though).   It has color-coded command messages to show that vs. data.